+++ src/4-executer/builtins/pwd.c
@@ -1,0 +1,28 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   pwd.c                                              :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: fschuber <fschuber@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/02/21 08:44:06 by fschuber          #+#    #+#             */
+/*   Updated: 2024/02/21 08:53:40 by fschuber         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../../../include/minishell.h"
+
+int execute_pwd()
+{
+	char *pwd;
+
+	pwd = getcwd(NULL, 0);
+	if (!pwd)
+	{
+		broadcast_builtin_error("pwd", -4, NULL);
+		return (errno);
+	}
+	ft_putendl_fd(pwd, STDOUT_FILENO);
+	free(pwd);
+	return (0);
+}