+++ src/input_loop.c
@@ -6,7 +6,8 @@ int run_crash_interface(t_program_data *program_data)
 
 	printf("\033[H\033[J\n%s\t ██████╗██████╗  █████╗ ███████╗██╗  ██╗\n\t%s██╔════╝██╔══██╗██╔══██╗██╔════╝██║  ██║\n\t%s██║     ██████╔╝███████║███████╗███████║\n\t%s██║     ██╔══██╗██╔══██║╚════██║██╔══██║\n\t%s╚██████╗██║  ██║██║  ██║███████║██║  ██║\n\t%s ╚═════╝╚═╝  ╚═╝╚═╝  ╚═╝╚══════╝╚═╝  ╚═╝%s\n\n", ANSI_COLOR_RED, ANSI_COLOR_GREEN, ANSI_COLOR_YELLOW, ANSI_COLOR_MAGENTA, ANSI_COLOR_CYAN, ANSI_COLOR_WHITE, ANSI_COLOR_RESET);
 	ft_printf("Welcome to %sCRASH%s, %s! - crazy robust & advanced shell!\n", ANSI_COLOR_RED, ANSI_COLOR_RESET, getenv("USER"));
-	ft_printf("              %sCreated by \033]8;;https://profile.intra.42.fr/users/fschuber\033\\fschuber\033]8;;\033\\ & \033]8;;https://profile.intra.42.fr/users/nburchha\033\\nburchha\033]8;;\033\\%s\n\n", ANSI_COLOR_CYAN, ANSI_COLOR_RESET);
+	ft_printf("          %sMinishell created by \033]8;;https://profile.intra.42.fr/users/fschuber\033\\fschuber\033]8;;\033\\ & \033]8;;https://profile.intra.42.fr/users/nburchha\033\\nburchha\033]8;;\033\\%s\n", ANSI_COLOR_CYAN, ANSI_COLOR_RESET);
+	ft_printf("\n                            ---\n\n");
 
 	(void) program_data;
 
@@ -20,6 +21,7 @@ int run_crash_interface(t_program_data *program_data)
 		if (input[0] == '\n')
 		{
 			cleanup(program_data->gc);
+			program_data->gc = createGarbageCollector();
 			continue;
 		}
 		input[ft_strlen(input) - 1] = '\0';